==> exploratory_notebooks/00_component_replacement_dictionary.ipynb <==
import json

major_positives = json.loads(open('major_positives.json').read())

sorted(major_positives)

unique_entries = sorted(list(set(major_positives)))
len(unique_entries)

unique_entries

unique_leading_words = sorted(list(set([x.split()[0] for x in major_positives])))
unique_leading_words

len(unique_leading_words)

new_gamer_synonyms = ['new gamer', 'non gamer', 'newcom', 'newbie', 'novice', 'rookie', 'starter', 'beginner']

# find all entries that contain a word from new_player_synonyms
new_player_entries = [x for x in unique_entries if any([y in x for y in new_gamer_synonyms])]
new_player_entries
