==> exploratory_notebooks/01_build_top_1000_generated_pros_cons.ipynb <==
import pandas as pd
from config import CONFIGS
import os
import re
import json

from utils.nlp_functions import initial_components_processing, build_sentiment_elements, filter_stopwords, strip_special_chars_and_lower
 
from typing import Tuple

import json

# hide warnings
import warnings
warnings.filterwarnings("ignore")

ENVIRONMENT = os.environ.get("ENVIRONMENT", "dev")
S3_SCRAPER_BUCKET = CONFIGS["s3_scraper_bucket"]
GAME_CONFIGS = CONFIGS["games"]
RATINGS_CONFIGS = CONFIGS["ratings"]
IS_LOCAL = True if os.environ.get("IS_LOCAL", "False").lower() == "true" else False

replacement_dictionary_path = 'data/prod/games/similarity_files/replacement_dictionary.json'
replacement_dictionary = json.load(open(replacement_dictionary_path))
replace_dict = replacement_dictionary['direct_replacements']
major_category_dict = replacement_dictionary['major_category_cleaning']
strip_list = replacement_dictionary['words_to_strip_all_right']

data_dir = 'data/prod/games/similarity_files/'

df = pd.read_pickle(f'{data_dir}top_1000_with_attached_rag.pkl')
desired_columns = ['BGGId','Name','Description']+[x for x in df.columns if 'generated' in x]

desc_df = df[desired_columns]
# desc_df.head()

desc_df['Description'] = desc_df['Description'].apply(lambda x: initial_components_processing(x, replace_dict))
desc_df['Description'] = desc_df['Description'].apply(filter_stopwords)
desc_df['Description'] = desc_df['Description'].apply(strip_special_chars_and_lower)

desc_df['generated_description'] = desc_df['generated_description'].apply(lambda x: initial_components_processing(x, replace_dict))
desc_df['generated_description'] = desc_df['generated_description'].apply(filter_stopwords)
desc_df['generated_description'] = desc_df['generated_description'].apply(strip_special_chars_and_lower)
desc_df.head()

components_positive = []
components_negative = []

components_positive.append(desc_df['generated_pros'].apply(lambda x: build_sentiment_elements(x, major_category_dict, strip_list)))
components_negative.append(desc_df['generated_cons'].apply(lambda x: build_sentiment_elements(x, major_category_dict, strip_list)))

desc_df['Positive_Components'] = [x[0] for x in components_positive[0]]
desc_df['Positive_Sentences']=  [x[1] for x in components_positive[0]]
desc_df['Negative_Components']= [x[0] for x in components_negative[0]]
desc_df['Negative_Sentences']=  [x[1] for x in components_negative[0]]

desc_df.head()

### Research problematic entries

problematic_entries = [373, 10, 136, 334, 214]
for problematic_entry in problematic_entries:
    print(f"\n\nEntry {problematic_entry}")
    print(desc_df.iloc[problematic_entry]['generated_pros'])
    print(desc_df.iloc[problematic_entry]['generated_cons'])
    print(desc_df.iloc[problematic_entry]['Positive_Components'])
    print(desc_df.iloc[problematic_entry]['Negative_Components'])
    print(desc_df.iloc[problematic_entry]['Positive_Sentences'])
    print(desc_df.iloc[problematic_entry]['Negative_Sentences'])
    

# find the entry with the most words in a component

max_len = 0
max_entry = ""

for i, row in desc_df.iterrows():
    row_sentences = row['Negative_Components']
    for sentence in row_sentences:
        if len(sentence.split()) > max_len:
            max_len = len(sentence.split())
            max_entry = i
            print(i, max_len, sentence)


### Save File

desc_df = desc_df.drop(columns=['Name','generated_pros','generated_cons'])
desc_df = desc_df.rename(columns={'generated_description':'About'})
desc_df.to_pickle(f"{data_dir}top_1000_cleaned_rag.pkl")
desc_df.head()

all_positive_components = []
all_negative_components = []

# raw_pro_components = desc_df['generated_pros'].apply(clean_field_to_integral_components)
# raw_negat_components = desc_df['generated_cons'].apply(clean_field_to_integral_components)

# for pos in raw_pro_components:
#     all_positive_components += pos

# for neg in raw_negat_components:
#     all_negative_components += neg

for pos in desc_df['Positive_Components']:
    all_positive_components += pos

for neg in desc_df['Negative_Components']:
    all_negative_components += neg

with open(f"{data_dir}major_positives.json", "w") as f:
    json.dump(all_positive_components, f)

with open (f"{data_dir}major_negatives.json", "w") as f:
    json.dump(all_negative_components, f)
